Update CampaignDetails.jsx
There was an error where if you clicked the donate button without inputing a number it would throw the loader forever. I fixed that by adding a conditional statement to check if the amount variable is an empty string before allowing the button to be clicked. I also added an alert using the window.alert() function.