Update CampaignDetails.jsx
There was an issue where if a user didnt connect their wallet and clicked donate. If would throw the loading screen forever. I have made an update to change the donate button to a connect button if the user does not have a connected wallet. And once it is connected, the button changes back into the donate button